+++ Source/Renderer/Public/DebugGui/DebugGuiHelper.h
@@ -97,7 +97,7 @@ namespace Renderer
 		RENDERER_API_EXPORT static void drawText(const char* text, float x, float y, bool drawBackground = true);
 		RENDERER_API_EXPORT static void drawGizmo(const CameraSceneItem& cameraSceneItem, GizmoSettings& gizmoSettings, Transform& transform);	// Using "ImGuizmo" ( https://github.com/CedricGuillemet/ImGuizmo )
 		RENDERER_API_EXPORT static void drawSkeleton(const CameraSceneItem& cameraSceneItem, const SkeletonMeshSceneItem& skeletonMeshSceneItem);
-		RENDERER_API_EXPORT static void drawGrid(const CameraSceneItem& cameraSceneItem, float cellSize = 1.0f, float yPosition = 0.0f);
+		RENDERER_API_EXPORT static void drawGrid(const CameraSceneItem& cameraSceneItem, float cellSize = 1.0f, double yPosition = 0.0f);
 
 
 	//[-------------------------------------------------------]
