+++ Source/Rhi/Private/NullRhi/NullRhi.cpp
@@ -72,7 +72,7 @@
 	#define NODEFERWINDOWPOS
 	#define NOMCX
 	#define NOCRYPT
-	#include <windows.h>
+	#include <Windows.h>
 
 	// Get rid of some nasty OS macros
 	#undef max
@@ -3540,7 +3540,7 @@ namespace
 		//[-------------------------------------------------------]
 		//[ Global definitions                                    ]
 		//[-------------------------------------------------------]
-		static constexpr Rhi::ImplementationDispatchFunction DISPATCH_FUNCTIONS[Rhi::CommandDispatchFunctionIndex::NUMBER_OF_FUNCTIONS] =
+		static constexpr Rhi::ImplementationDispatchFunction DISPATCH_FUNCTIONS[static_cast<uint8_t>(Rhi::CommandDispatchFunctionIndex::NUMBER_OF_FUNCTIONS)] =
 		{
 			// Command buffer
 			&ImplementationDispatch::ExecuteCommandBuffer,
